Move clipboard selection builder to browser layer to fix layering violation
common/ cannot import from browser/. Move buildTableSelectionFromClipboardData
to browser/utils.ts and revert common/utils.ts to its original state.
